fix(sabnzbd ): correct probes (#36184)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->
⚒️ Fixes  # <!--(issue)-->

SABnzbd have a huge log spam, see hereunder this will be a lot of times
and each few minutes.
Seems to be the probes.
This will fix it and logs are nice and quiet now.

```
2025-06-10 17:09:18.173 | Exception ignored in: <function IOBase.__del__ at 0x7fe09dc0cb80> |  
-- | -- | --
  |   | 2025-06-10 17:09:18.173 | Traceback (most recent call last): |  
  |   | 2025-06-10 17:09:18.173 | File "/usr/local/lib/python3.13/_pyio.py", line 418, in __del__ |  
  |   | 2025-06-10 17:09:18.173 | self.close() |  
  |   | 2025-06-10 17:09:18.173 | File "/usr/local/lib/python3.13/_pyio.py", line 1313, in close |  
  |   | 2025-06-10 17:09:18.173 | self.flush() |  
  |   | 2025-06-10 17:09:18.173 | File "/usr/local/lib/python3.13/_pyio.py", line 1274, in flush |  
  |   | 2025-06-10 17:09:18.174 | self._flush_unlocked() |  
  |   | 2025-06-10 17:09:18.174 | File "/usr/local/lib/python3.13/site-packages/cheroot/makefile.py", line 33, in _flush_unlocked |  
  |   | 2025-06-10 17:09:18.174 | n = self.raw.write(bytes(self._write_buf)) |  
  |   | 2025-06-10 17:09:18.174 | File "/usr/local/lib/python3.13/socket.py", line 737, in write |  
  |   | 2025-06-10 17:09:18.174 | return self._sock.send(b) |  
  |   | 2025-06-10 17:09:18.174 | OSError: [Errno 9] Bad file descriptor
```

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code
- [ ] 📜 Documentation Changes

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->
On my own cluster:

https://github.com/Boemeltrein/TalosCluster/commit/5b6897e43b3c81ddd9f1ce07a2926f90ae4f3174

**📃 Notes:**
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made changes to the documentation
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning
- [x] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):`, `chore(chart-name):`, `docs(chart-name):` or
`fix(docs):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._