feat(openwebrxplus): Add OpenWebRX+ chart (#36626)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->
Adds Helm chart for [OpenWebRX+](https://fms.komkon.org/OWRX/)

**⚙️ Type of change**

- [x] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code
- [ ] 📜 Documentation Changes

**🧪 How Has This Been Tested?**

- Currently running it via Flux to a K3s cluster on NixOS, from this
branch.
- Service comes up without issues.
- Hardware SDR module IO runs fine when `persistence.usb` is `enabled:
true`. Having said that, ensuring that USB paths and permissions match
what the `openwebrx` process expects is left to the end user. That's
because the necessary configuration will differ, depending on the
environment. While running as `privileged` would resolve permission
issues, I believe it's best to make it the user's choice, and not a
silent default.


![image](https://github.com/user-attachments/assets/01a07ebd-0e27-4d9c-a79b-a3f8a4f0b575)


**📃 Notes:**

- Nice to have in subsequent PRs:
- declarative openwebrx configuration. Declared in the helm chart
(plaintext), and persisted as a ConfigMap which is mounted into the
container;

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [x] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made changes to the documentation
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [ ] ⬆️ I increased versions for any altered app according to semantic
versioning
- [x] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):`, `chore(chart-name):`, `docs(chart-name):` or
`fix(docs):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [x] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Alfred Göppel <43101280+alfi0812@users.noreply.github.com>
Co-authored-by: Alfred Göppel <43101280+alfi0812@users.noreply.github.com>